Basic set of webpack configs.
